requirements.txt
python-telegram-bot
python-dotenv
pymongo
motor
tgcrypto
aiohttp
loguru
pytz
aiofiles
requests